backend/requirements.txt
flask
pytesseract
Pillow
opencv-python
python-dotenv
google-generativeai
utils
config
requests
numpy
flask-cors